requirements.txt
Flask==0.10.1
Flask-Cors==2.1.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-Testing==0.4.2
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
SQLAlchemy==1.0.10
Werkzeug==0.11.2
aniso8601==1.1.0
argparse==1.4.0
blinker==1.4
colorama==0.3.5
cov-core==1.15.0
coverage==4.0.3
decorator==4.0.6
factory-boy==2.6.0
fake-factory==0.5.3
gunicorn==19.4.1
ipython==4.0.1
ipython-genutils==0.1.0
itsdangerous==0.24
logutils==0.3.3
marshmallow==2.4.2
newrelic==2.60.0.46
nose2==0.5.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
pycrypto==2.6.1
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.0
simplegeneric==0.8.1
simplejson==3.8.1
six==1.10.0
traitlets==4.0.0
